requirements.txt
Flask>=2.3
Flask-Login
Flask-Mail
Flask-WTF
Flask-Migrate
Flask-Limiter
Flask-Talisman
Flask-Bcrypt
Flask-SQLAlchemy
python-dotenv
itsdangerous
pymysql
email-validator
pyotp
qrcode
Pillow
pytz
requests
Flask-SocketIO
authlib
cryptography
gunicorn
eventlet>=0.33.3
python-socketio[client]>=5.9.0
redis>=4.5.0
greenlet>=2.0.0